RavenDB-18646 implemented /databases/*/debug/queries/kill endpoint
