+++ src/Raven.Server/Documents/Queries/Dynamic/DynamicQueryRunner.cs
@@ -20,7 +20,7 @@ using PatchRequest = Raven.Server.Documents.Patch.PatchRequest;
 
 namespace Raven.Server.Documents.Queries.Dynamic
 {
-    public class DynamicQueryRunner : AbstractQueryRunner
+    public class DynamicQueryRunner : AbstractDatabaseQueryRunner
     {
         private readonly IndexStore _indexStore;
