+++ src/Raven.Server/Documents/Queries/StaticIndexQueryRunner.cs
@@ -15,7 +15,7 @@ using PatchRequest = Raven.Server.Documents.Patch.PatchRequest;
 
 namespace Raven.Server.Documents.Queries
 {
-    public class StaticIndexQueryRunner : AbstractQueryRunner
+    public class StaticIndexQueryRunner : AbstractDatabaseQueryRunner
     {
         public StaticIndexQueryRunner(DocumentDatabase database) : base(database)
         {
