+++ src/Raven.Server/Documents/Sharding/Queries/AbstractShardedQueryProcessor.cs
@@ -19,7 +19,8 @@ using Voron;
 
 namespace Raven.Server.Documents.Sharding.Queries;
 
-public abstract class AbstractShardedQueryProcessor<TCommand, TResult, TCombinedResult> : IDisposable where TCommand : RavenCommand<TResult>
+public abstract class AbstractShardedQueryProcessor<TCommand, TResult, TCombinedResult> : IDisposable 
+    where TCommand : RavenCommand<TResult>
 {
     const string LimitToken = "__raven_limit";
