+++ src/Raven.Server/Documents/Sharding/Queries/ShardedQueryRunner.cs
@@ -1,0 +1,7 @@
+﻿using Raven.Server.Documents.Queries;
+
+namespace Raven.Server.Documents.Sharding.Queries;
+
+public class ShardedQueryRunner : AbstractQueryRunner
+{
+}
