+++ src/Raven.Server/Documents/Sharding/ShardedDatabaseContext.Queries.cs
@@ -1,0 +1,8 @@
+﻿using Raven.Server.Documents.Sharding.Queries;
+
+namespace Raven.Server.Documents.Sharding;
+
+public partial class ShardedDatabaseContext
+{
+    public ShardedQueryRunner QueryRunner;
+}
